Move DaemonThreadFactory to internal package.
